the -M is a double, but the setOptions method used "Integer.parseInt" - changed that to "Double.parseDouble" (reason: crashing the XML-Deserialization)
git-svn-id: https://svn.cms.waikato.ac.nz/svn/weka/branches/book2ndEd-branch@2390 e0a1b77d-ad91-4216-81b1-defd5f83fa92

+++ weka/classifiers/trees/RandomTree.java
@@ -32,7 +32,7 @@ import java.util.*;
  *
  * @author Eibe Frank (eibe@cs.waikato.ac.nz)
  * @author Richard Kirkby (rkirkby@cs.waikato.ac.nz)
- * @version $Revision: 1.8 $
+ * @version $Revision: 1.8.2.1 $
  */
 public class RandomTree extends Classifier 
   implements OptionHandler, WeightedInstancesHandler, Randomizable {
@@ -261,7 +261,7 @@ public class RandomTree extends Classifier
     }
     String minNumString = Utils.getOption('M', options);
     if (minNumString.length() != 0) {
-      m_MinNum = (double)Integer.parseInt(minNumString);
+      m_MinNum = Double.parseDouble(minNumString);
     } else {
       m_MinNum = 1;
     }
